+++ packages/framework/src/Framework/Actions/PaginatesPublicationListing.php
@@ -5,6 +5,12 @@ declare(strict_types=1);
 namespace Hyde\Framework\Actions;
 
 use Hyde\Framework\Features\Publications\Models\PublicationType;
+use Hyde\Framework\Features\Publications\PublicationService;
+use Hyde\Hyde;
+use Hyde\Pages\VirtualPage;
+
+use function array_merge;
+use function file_put_contents;
 
 class PaginatesPublicationListing
 {
@@ -18,5 +24,52 @@ class PaginatesPublicationListing
     public function __invoke(): array
     {
         return [];
+    }
+
+
+    protected function generatePaginationPages(): void
+    {
+        $pubType = $this->page->type;
+        $pages = PublicationService::getPublicationsForPubType($pubType);
+
+        $count = 25; //FIXme get form type
+        $chunks = $pages->chunk($count);
+
+        foreach ($chunks as $index => $chunk) {
+            $page = $index + 1;
+            $data = [
+                'publications' => $chunk,
+                'pagination' => [
+                    'current' => $page,
+                    'total' => $chunks->count(),
+                    'next' => $page < $chunks->count() ? $page + 1 : null,
+                    'previous' => $page > 1 ? $page - 1 : null,
+                    'offset' => $index * $count + 1,
+                ],
+            ];
+
+            $this->savePaginationPage($pubType, $page, $data);
+        }
+    }
+
+    protected function savePaginationPage(PublicationType $pubType, int $pageNumber, array $data)
+    {
+        $identifier = "{$pubType->getDirectory()}/page-$pageNumber";
+
+        $path = "$identifier.html";
+        $data = array_merge([
+            'type' => $pubType,
+            'paginator' => (object) $data['pagination'],
+        ], $data);
+
+        $page = new VirtualPage($identifier, array_merge([
+            'title' => $pubType->name." (Page - $pageNumber)",
+        ], $data), '', 'hyde::layouts.publication_paginated_list');
+
+        file_put_contents(Hyde::sitePath($path), $page->compile());
+        if ($pageNumber === 1) { //fixme this overwrites the other page leading to wasted compilation time and also makes us lose that page's data // we also prob dont want to show the number in the title
+            $path = "{$pubType->getDirectory()}/index.html";
+            file_put_contents(Hyde::sitePath($path), $page->compile());
+        }
     }
 }
